make particle ids signed

+++ include/LHAPDF/LogBicubicInterpolator.h
@@ -19,7 +19,7 @@ namespace LHAPDF {
   public:
 
     /// Implementation of (x,Q2) interpolation
-    double _interpolateXQ2(const KnotArray& subgrid, double x, size_t ix, double q2, size_t iq2, size_t id) const;
+    double _interpolateXQ2(const KnotArray& subgrid, double x, size_t ix, double q2, size_t iq2, int id) const;
 
     /// @brief A single set of cached x-variables
     struct XCache {
